Revert "header outproduct define error"
This reverts commit 9f2fa6af6dc0ee1095225041d8e5cc87bdc7ac3c.

+++ lib/TH/vector/AVX512.h
@@ -21,5 +21,5 @@ void THFloatVector_cadd_AVX512(float *z, const float *x, const float *y, const f
 void THFloatVector_adds_AVX512(float *y, const float *x, const float c, const ptrdiff_t n);
 
 void THDoubleVector_outerproduct_AVX512(double *z, const double *x, const double *y, const ptrdiff_t m, const ptrdiff_t n);
-void THFloatVector_outerproduct_AVX512(float *z, const float *x, const float *y, const ptrdiff_t m, const ptrdiff_t n);
+void THFloatVector_outerproduct_AVX512(float *z, const double *x, const double *y, const ptrdiff_t m, const ptrdiff_t n);
 #endif
